requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==13.1
pipecat-ai==0.0.46
python-dotenv==1.0.1
